requirements/common.txt
Django==1.9.11
djangorestframework==3.3.3
psycopg2==2.6.1
djoser==0.4.3
django-allauth==0.25.2
django-cors-headers==1.1.0
django-filter==0.12.0
django-crispy-forms==1.6.0
django-parsley==0.6
django-formtools==1.0
django-countries==3.4.1
django-leaflet==0.18.0
django-sass-processor==0.3.4
djangorestframework-gis==0.10.1
jsonschema==2.5.1
rfc3987==1.3.5
drfdocs-cadasta==0.0.12
django-tutelary==0.1.21
django-audit-log==0.7.0
django-simple-history==1.8.1
simplejson==3.8.1
django-widget-tweaks==1.4.1
django-buckets==0.1.18
pyxform-cadasta==0.9.22
python-magic==0.4.11
Pillow==3.2.0
django-jsonattrs==0.1.21
openpyxl==2.3.5
pytz==2016.4
shapely==1.5.16
gdal==1.10.0
pylibmc==1.5.1
awscli==1.11.23
pandas==0.19.0
requests==2.11.1
